Split out Disassembler into modules that define a generic disassembler and then processor specific modules.
Current prossesor suport modules are for 6502 and 6809.
Fixed several bugs in the symbol management routines that did not affect the
6502 module but caused problems for the 6809 module.

Changed the way that the CPU command works in the control file, to be more
friendly, CPU is now specified by name : 6502, 65c02, wd6502, 6512, 6809.

Needs regression testing, needs more 6809 testing, but was able to successfully
generate a source file that could be re-assembled and be identical to original.